abstract ui and tasklist
