+++ src/main/data/duke.txt
@@ -1,3 +1,1 @@
 T ~   ~  borrow ~ 
-D ~   ~  return book ~ 2022-01-23
-T ~   ~  return book ~
